+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.0.1 under development
 
-- no changes in this release.
+- Bug #51: Fix solution description to `WidgetFactoryInitializationException` (devanych)
 
 ## 1.0.0 December 12, 2021